restaurant_finder/apple_tools/requirements.txt
fastmcp>=0.1.0
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
starlette>=0.27.0
